requirement.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor-5==0.2.13
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.2
mysql==0.0.3
mysqlclient==2.2.4
pillow==10.3.0
sqlparse==0.5.0
tzdata==2024.1
